+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Hue and Cry button added to combat tracker.
+- Foundry v11 api support.
 ### Changed
 - Removed extraneous details from combat tracker config.
+- Reach animation tweaks.
 
 ## [0.3.25] - 2023-05-21
 ### Added